Rename src/client to src/browser
closes este/este/#559

Given the native app is also a client, it seems to make more sense to call the
"client" folder "browser" since it holds views and styles specific to the
browser application.
