+++ src/browser/.eslintrc
