+++ src/browser/app/Footer.react.js
