+++ src/browser/auth/__test__/Login.spec.js
