+++ src/browser/lib/focusInvalidField.js
