+++ src/browser/lib/logRenderTime.js
