+++ src/browser/main.js
