+++ src/server/frontend/render.js
@@ -5,7 +5,7 @@ import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 import config from '../config';
 import configureStore from '../../common/configureStore';
-import createRoutes from '../../client/createRoutes';
+import createRoutes from '../../browser/createRoutes';
 import serialize from 'serialize-javascript';
 import useragent from 'useragent';
 import {HOT_RELOAD_PORT} from '../../../webpack/constants';
